Moved restock state management to mock
