+++ src/api/mock/restock/utils.ts
@@ -1,0 +1,5 @@
+import {dummyRestockData} from './data';
+
+export const mockFetchRestockData = async () => {
+  return dummyRestockData;
+};
