+++ src/components/elements/table/PurchaseModal.tsx
@@ -6,7 +6,7 @@ import DialogActions from '@material-ui/core/DialogActions';
 import DialogContent from '@material-ui/core/DialogContent';
 import DialogTitle from '@material-ui/core/DialogTitle';
 
-import {ItemDataInfo} from '../../../state/restock/restockData';
+import {ItemDataInfo} from '../../../api/definitions/restock/data';
 import UIButton from '../ui/Button';
 import PurchaseOverview from './PurchaseOverview';
