+++ src/components/elements/table/RestockPurchasesTable.tsx
@@ -1,4 +1,4 @@
-import React, {useEffect} from 'react';
+import React from 'react';
 
 import Paper from '@material-ui/core/Paper';
 import {createStyles, makeStyles, Theme} from '@material-ui/core/styles';
@@ -8,12 +8,12 @@ import TableCell from '@material-ui/core/TableCell';
 import TableContainer from '@material-ui/core/TableContainer';
 import TablePagination from '@material-ui/core/TablePagination';
 import TableRow from '@material-ui/core/TableRow';
-import {unwrapResult} from '@reduxjs/toolkit';
 
+import {FetchStatus} from '../../../api/definitions/misc';
+import {RestockData, RestockItemInfo} from '../../../api/definitions/restock/data';
+import {RestockStatusType} from '../../../api/definitions/restock/data';
+import {mockFetchRestockData} from '../../../api/mock/restock/utils';
 import {alertDispatchers} from '../../../state/alert/dispatchers';
-import {RestockData, RestockItemInfo} from '../../../state/restock/restockData';
-import {RestockStatusType} from '../../../state/restock/restockData';
-import {restockDispatchers} from '../../../state/restock/restockDispatchers';
 import {useDispatch} from '../../../state/store';
 import {Order, sort} from '../../../utils/sort';
 import {NoData} from './NoData';
@@ -55,7 +55,10 @@ type PageState = {
 const RestockPurchasesTable = () => {
   const classes = useStyles();
   const dispatch = useDispatch();
-  const [restockData, setRestockData] = React.useState<RestockData>([]);
+  const [restockData, setRestockData] = React.useState<FetchStatus<RestockData>>({
+    fetched: false,
+    fetching: false,
+  });
   const [pageState, setPageState] = React.useState<PageState>({
     page: 0,
     rowsPerPage: 10,
@@ -63,18 +66,26 @@ const RestockPurchasesTable = () => {
     orderBy: 'id',
   });
 
-  const getData = () => {
-    dispatch(restockDispatchers.fetchRestockData())
-      .then(unwrapResult)
-      .then((result) => setRestockData(result))
+  if (!restockData.fetched) {
+    mockFetchRestockData()
+      .then((data) => {
+        setRestockData({
+          ...restockData,
+          fetched: true,
+          fetching: false,
+          response: data,
+        });
+      })
       .catch((error) => {
+        setRestockData({
+          ...restockData,
+          fetched: true,
+          fetching: false,
+          error: error.message,
+        });
         dispatch(alertDispatchers.showAlert({severity: 'error', message: error.message}));
       });
-  };
-
-  useEffect(() => {
-    getData();
-  }, []);
+  }
 
   const handleRequestSort = (event: React.MouseEvent<HTMLSpanElement>, property: keyof RestockItemInfo) => {
     const isAsc = pageState.orderBy === property && pageState.order === 'asc';
@@ -97,12 +108,14 @@ const RestockPurchasesTable = () => {
     });
   };
 
-  if (!restockData.length) {
+  if (!restockData.response?.length) {
     return <NoData/>;
   }
 
-  const emptyRows = pageState.rowsPerPage - Math.min(pageState.rowsPerPage,
-    restockData.length - pageState.page * pageState.rowsPerPage);
+  const emptyRows = pageState.rowsPerPage - Math.min(
+    pageState.rowsPerPage,
+    restockData.response.length - pageState.page * pageState.rowsPerPage,
+  );
 
   return (
     <div className={classes.root}>
@@ -119,7 +132,7 @@ const RestockPurchasesTable = () => {
               onRequestSort={handleRequestSort}
             />
             <TableBody>
-              {sort(restockData, {
+              {sort(restockData.response, {
                 order: pageState.order,
                 sortKey: (item) => item[pageState.orderBy],
               })
@@ -128,27 +141,29 @@ const RestockPurchasesTable = () => {
                   pageState.page * pageState.rowsPerPage + pageState.rowsPerPage,
                 )
                 .map((row) => {
-                  return <TableRow
-                    hover
-                    tabIndex={-1}
-                    key={row.id}
-                  >
-                    <TableCell component="th" scope="row">
-                      {row.id}
-                    </TableCell>
-                    <TableCell>{row.purchaseDate}</TableCell>
-                    <TableCell>
-                      <Status
-                        color={statusColors[row.status]}
-                      />
-                      {row.status}
-                    </TableCell>
-                    <TableCell>{row.type}</TableCell>
-                    <TableCell>{row.totalPrice}</TableCell>
-                    <TableCell>
-                      <PurchaseModal data={row.itemsData} orderId={row.id}/>
-                    </TableCell>
-                  </TableRow>;
+                  return (
+                    <TableRow
+                      hover
+                      tabIndex={-1}
+                      key={row.id}
+                    >
+                      <TableCell component="th" scope="row">
+                        {row.id}
+                      </TableCell>
+                      <TableCell>{row.purchaseDate}</TableCell>
+                      <TableCell>
+                        <Status
+                          color={statusColors[row.status]}
+                        />
+                        {row.status}
+                      </TableCell>
+                      <TableCell>{row.type}</TableCell>
+                      <TableCell>{row.totalPrice}</TableCell>
+                      <TableCell>
+                        <PurchaseModal data={row.itemsData} orderId={row.id}/>
+                      </TableCell>
+                    </TableRow>
+                  );
                 })}
               {emptyRows > 0 && <TableRow style={{height: 53 * emptyRows}}>
                 <TableCell colSpan={6}/>
@@ -159,7 +174,7 @@ const RestockPurchasesTable = () => {
         <TablePagination
           rowsPerPageOptions={[10, 25, 100]}
           component="div"
-          count={restockData.length}
+          count={restockData.response.length}
           rowsPerPage={pageState.rowsPerPage}
           page={pageState.page}
           onChangePage={handleChangePage}
