+++ src/state/reducer.ts
@@ -4,7 +4,6 @@ import storage from 'redux-persist/lib/storage';
 
 import alertReducer from './alert/reducer';
 import authReducer from './auth/reducer';
-import restockReducer from './restock/restockReducer';
 
 const persistConfig = {
   key: 'root',
@@ -16,7 +15,6 @@ const rootReducer = combineReducers(
   {
     auth: authReducer,
     alert: alertReducer,
-    restock: restockReducer,
   },
 );
