--- src/state/restock/restockName.ts
@@ -1,6 +1,0 @@
-export const RESTOCK_STATE_NAME = 'restock';
-
-export enum RestockDispatcherName {
-    UPDATE_RESTOCK_SETTINGS = 'updateRestockSettings',
-    FETCH_RESTOCK_DATA = 'fetchRestockData',
-}
