--- src/state/restock/restockReducer.ts
@@ -1,27 +1,0 @@
-import {createSlice} from '@reduxjs/toolkit';
-
-import {restockDispatchers} from './restockDispatchers';
-import {RESTOCK_STATE_NAME} from './restockName';
-import {RestockState} from './restockState';
-
-const initialState: RestockState = {
-  restockData: [],
-};
-
-const restockSlice = createSlice({
-  name: RESTOCK_STATE_NAME,
-  initialState,
-  reducers: {},
-  extraReducers: (builder) => {
-    [restockDispatchers.fetchRestockData.fulfilled].forEach((thunkCase) => {
-      builder.addCase(
-        thunkCase,
-        (state, {payload}) => {
-          state.restockData = payload;
-        },
-      );
-    });
-  },
-});
-
-export default restockSlice.reducer;
