--- src/state/restock/restockSelector.ts
@@ -1,7 +1,0 @@
-import {useSelector} from 'react-redux';
-
-import {ReduxState} from '../state';
-
-export const useRestockSelector = () => {
-  return useSelector((state: ReduxState) => state.restock);
-};
