--- src/state/restock/restockState.ts
@@ -1,6 +1,0 @@
-import {StateBase} from '../base';
-import {RestockData} from './restockData';
-
-export type RestockState = StateBase & {
-    restockData: RestockData;
-}
